Move Channel filter and EnabledFilter
